# backend/tests/snapshots/openapi_schema.json
{
  "components": {
    "schemas": {
...
      "BuscaRequest": {
        "description": "Request schema for /buscar endpoint.\n\nValidates user input for procurement search:\n- At least 1 Brazilian state (UF) must be selected\n- Dates must be in YYYY-MM-DD format\n- data_inicial must be <= data_final\n- Date range cannot exceed 30 days\n- data_final cannot be in the future\n- valor_maximo must be >= valor_minimo (if both provided)\n\nExamples:\n    >>> request = BuscaRequest(\n    ...     ufs=[\"SP\", \"RJ\"],\n    ...     data_inicial=\"2025-01-01\",\n    ...     data_final=\"2025-01-31\"\n    ... )\n    >>> request.ufs\n    ['SP', 'RJ']\n\n    >>> # With new P0/P1 filters\n    >>> request = BuscaRequest(\n    ...     ufs=[\"SP\"],\n    ...     data_inicial=\"2025-01-01\",\n    ...     data_final=\"2025-01-31\",\n    ...     status=StatusLicitacao.RECEBENDO_PROPOSTA,\n    ...     modalidades=[1, 2, 6],\n    ...     valor_minimo=50000,\n    ...     valor_maximo=500000,\n    ...     esferas=[EsferaGovernamental.MUNICIPAL]\n    ... )",
        "example": {
          "data_final": "2025-01-31",
          "data_inicial": "2025-01-01",
          "esferas": [
            "M"
          ],
          "modalidades": [
            1,
            2,
            6
          ],
          "ordenacao": "data_desc",
          "status": "recebendo_proposta",
          "ufs": [
            "SP",
            "RJ"
          ],
          "valor_maximo": 500000,
          "valor_minimo": 50000
        },
        "properties": {
          "data_final": {
            "description": "End date in YYYY-MM-DD format",
            "examples": [
              "2025-01-31"
            ],
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "title": "Data Final",
            "type": "string"
          },
          "data_inicial": {
            "description": "Start date in YYYY-MM-DD format",
            "examples": [
              "2025-01-01"
            ],
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "title": "Data Inicial",
            "type": "string"
          },
          "esferas": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/EsferaGovernamental"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Lista de esferas governamentais ('F'=Federal, 'E'=Estadual, 'M'=Municipal). None = todas as esferas.",
            "examples": [
              [
                "M",
                "E"
              ]
            ],
            "title": "Esferas"
          },
          "exclusion_terms": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Custom exclusion terms. Overrides sector exclusions when provided.",
            "title": "Exclusion Terms"
          },
          "itens_por_pagina": {
            "default": 20,
            "description": "Quantidade de itens por p\u00e1gina. Valores permitidos: 10, 20, 50, 100. Padr\u00e3o: 20.",
            "examples": [
              10,
              20,
              50,
              100
            ],
            "maximum": 100.0,
            "minimum": 10.0,
            "title": "Itens Por Pagina",
            "type": "integer"
          },
          "modalidades": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Lista de c\u00f3digos de modalidade conforme Lei 14.133/2021. C\u00f3digos v\u00e1lidos: 1 (Preg\u00e3o Eletr\u00f4nico), 2 (Preg\u00e3o Presencial), 3 (Concorr\u00eancia), 6 (Dispensa), 7 (Inexigibilidade), 9 (Leil\u00e3o), 10 (Di\u00e1logo Competitivo), 11 (Concurso). None = todas as modalidades. NOTA: C\u00f3digos 4 (Tomada de Pre\u00e7os) e 5 (Convite) foram removidos por serem da Lei 8.666/93 (revogada).",
            "examples": [
              [
                1,
                2,
                6
              ]
            ],
            "title": "Modalidades"
          },
          "municipios": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Lista de c\u00f3digos IBGE de munic\u00edpios para filtrar. None = todos os munic\u00edpios das UFs selecionadas.",
            "examples": [
              [
                "3550308",
                "3304557"
              ]
            ],
            "title": "Municipios"
          },
          "ordenacao": {
            "default": "data_desc",
            "description": "Crit\u00e9rio de ordena\u00e7\u00e3o dos resultados: 'data_desc' (mais recente), 'data_asc' (mais antigo), 'valor_desc' (maior valor), 'valor_asc' (menor valor), 'prazo_asc' (prazo mais pr\u00f3ximo), 'relevancia' (score de matching).",
            "examples": [
              "data_desc",
              "valor_desc",
              "prazo_asc"
            ],
            "title": "Ordenacao",
            "type": "string"
          },
          "orgaos": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Lista de nomes de \u00f3rg\u00e3os/entidades para filtrar (busca parcial). None = todos os \u00f3rg\u00e3os.",
            "examples": [
              [
                "Prefeitura de S\u00e3o Paulo",
                "Minist\u00e9rio da Sa\u00fade"
              ]
            ],
            "title": "Orgaos"
          },
          "pagina": {
            "default": 1,
            "description": "N\u00famero da p\u00e1gina de resultados (1-indexed). Padr\u00e3o: 1.",
            "examples": [
              1,
              2,
              3
            ],
            "minimum": 1.0,
            "title": "Pagina",
            "type": "integer"
          },
          "search_id": {
            "anyOf": [
              {
                "maxLength": 36,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Client-generated UUID for SSE progress tracking via /buscar-progress/{search_id}.",
            "title": "Search Id"
          },
          "setor_id": {
            "default": "vestuario",
            "description": "Sector ID for keyword filtering (e.g., 'vestuario', 'alimentos', 'informatica')",
            "examples": [
              "vestuario"
            ],
            "title": "Setor Id",
            "type": "string"
          },
          "show_all_matches": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": false,
            "description": "When True, bypasses minimum match floor and returns all results with at least 1 keyword match (for 'show hidden results' feature).",
            "title": "Show All Matches"
          },
          "status": {
            "$ref": "#/components/schemas/StatusLicitacao",
            "default": "todos",
            "description": "Status da licita\u00e7\u00e3o para filtrar. Padr\u00e3o: 'todos' (sem filtro de status). IMPORTANTE: Filtro de status desabilitado por padr\u00e3o devido a valores inconsistentes na API PNCP. Use 'todos' para m\u00e1xima cobertura de resultados.",
            "examples": [
              "todos",
              "recebendo_proposta",
              "em_julgamento",
              "encerrada"
            ]
          },
          "termos_busca": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Custom search terms. Use commas to separate multi-word phrases (e.g., 'levantamento topogr\u00e1fico, terraplenagem, drenagem'). Without commas, spaces separate individual keywords (legacy mode).",
            "examples": [
              "jaleco avental",
              "levantamento topogr\u00e1fico, terraplenagem, drenagem"
            ],
            "title": "Termos Busca"
          },
          "ufs": {
            "description": "List of Brazilian state codes (e.g., ['SP', 'RJ', 'MG'])",
            "examples": [
              [
                "SP",
                "RJ"
              ]
            ],
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Ufs",
            "type": "array"
          },
          "valor_maximo": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Valor m\u00e1ximo estimado da licita\u00e7\u00e3o em BRL. None = sem limite superior.",
            "examples": [
              5000000.0
            ],
            "title": "Valor Maximo"
          },
          "valor_minimo": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Valor m\u00ednimo estimado da licita\u00e7\u00e3o em BRL. None = sem limite inferior.",
            "examples": [
              50000.0
            ],
            "title": "Valor Minimo"
          }
        },
        "required": [
          "ufs",
          "data_inicial",
          "data_final"
        ],
        "title": "BuscaRequest",
        "type": "object"
      },
      "BuscaResponse": {
        "description": "Response schema for /buscar endpoint.\n\nReturns the complete search results including:\n- AI-generated executive summary\n- Excel file as base64-encoded string (optional, based on plan)\n- Statistics about raw vs filtered results\n- Quota information for user awareness\n\nThe Excel file can be decoded and downloaded by the frontend.",
        "example": {
          "excel_available": true,
          "excel_base64": "UEsDBBQABg...",
          "quota_remaining": 26,
          "quota_used": 24,
          "resumo": {
            "destaques": [
              "3 urgentes"
            ],
            "resumo_executivo": "Encontradas 15 licita\u00e7\u00f5es.",
            "total_oportunidades": 15,
            "valor_total": 2300000.0
          },
          "total_filtrado": 15,
          "total_raw": 523,
          "ultima_atualizacao": "2026-02-10T14:30:00Z"
        },
        "properties": {
          "download_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Signed URL for direct Excel download (60min TTL, preferred over base64)",
            "title": "Download Url"
          },
          "excel_available": {
            "description": "Whether Excel export is available for user's plan",
            "title": "Excel Available",
            "type": "boolean"
          },
          "excel_base64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Excel file encoded as base64 string (None if plan doesn't allow or storage used)",
            "title": "Excel Base64"
          },
          "filter_relaxed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "True if the minimum match filter was relaxed from strict to 1 due to zero results",
            "title": "Filter Relaxed"
          },
          "filter_stats": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FilterStats"
              },
              {
                "type": "null"
              }
            ],
            "description": "Breakdown of filter rejection reasons"
          },
          "hidden_by_min_match": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Number of bids that matched at least 1 term but were below the minimum match floor",
            "title": "Hidden By Min Match"
          },
          "licitacoes": {
            "description": "List of individual bids for display. FREE tier shows 5-10 fully, rest blurred.",
            "items": {
              "$ref": "#/components/schemas/LicitacaoItem"
            },
            "title": "Licitacoes",
            "type": "array"
          },
          "quota_remaining": {
            "description": "Monthly searches remaining",
            "minimum": 0.0,
            "title": "Quota Remaining",
            "type": "integer"
          },
          "quota_used": {
            "description": "Monthly searches used after this request",
            "minimum": 0.0,
            "title": "Quota Used",
            "type": "integer"
          },
          "resumo": {
            "$ref": "#/components/schemas/ResumoLicitacoes",
            "description": "Executive summary (AI-generated or fallback)"
          },
          "source_stats": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Per-source fetch metrics when multi-source is active",
            "title": "Source Stats"
          },
          "stopwords_removidas": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Stopwords stripped from user input (for transparency)",
            "title": "Stopwords Removidas"
          },
          "termos_utilizados": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Keywords effectively used for filtering (after stopword removal)",
            "title": "Termos Utilizados"
          },
          "total_filtrado": {
            "description": "Records after applying filters (UF, value, keywords, deadline)",
            "minimum": 0.0,
            "title": "Total Filtrado",
            "type": "integer"
          },
          "total_raw": {
            "description": "Total records fetched from PNCP API (before filtering)",
            "minimum": 0.0,
            "title": "Total Raw",
            "type": "integer"
          },
          "ultima_atualizacao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ISO timestamp of when search results were generated",
            "title": "Ultima Atualizacao"
          },
          "upgrade_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Message shown when Excel is blocked, encouraging upgrade",
            "title": "Upgrade Message"
          }
        },
        "required": [
          "resumo",
          "excel_available",
          "quota_used",
          "quota_remaining",
          "total_raw",
          "total_filtrado"
        ],
        "title": "BuscaResponse",
        "type": "object"
      },
...
      "EsferaGovernamental": {
        "description": "Esfera governamental para filtro de busca.\n\nDefine o n\u00edvel de governo do \u00f3rg\u00e3o contratante:\n- FEDERAL (F): Uni\u00e3o, minist\u00e9rios, autarquias federais\n- ESTADUAL (E): Estados, DF, secretarias estaduais\n- MUNICIPAL (M): Prefeituras, c\u00e2maras municipais",
        "enum": [
          "F",
          "E",
          "M"
        ],
        "title": "EsferaGovernamental",
        "type": "string"
      },
...
      "FilterStats": {
        "description": "Statistics about filter rejection reasons.",
        "properties": {
          "rejeitadas_keyword": {
            "default": 0,
            "description": "Rejected by keyword match (zero matches)",
            "title": "Rejeitadas Keyword",
            "type": "integer"
          },
          "rejeitadas_min_match": {
            "default": 0,
            "description": "Rejected by minimum match floor (had matches but below threshold)",
            "title": "Rejeitadas Min Match",
            "type": "integer"
          },
          "rejeitadas_outros": {
            "default": 0,
            "description": "Rejected by other reasons",
            "title": "Rejeitadas Outros",
            "type": "integer"
          },
          "rejeitadas_prazo": {
            "default": 0,
            "description": "Rejected by deadline",
            "title": "Rejeitadas Prazo",
            "type": "integer"
          },
          "rejeitadas_uf": {
            "default": 0,
            "description": "Rejected by UF filter",
            "title": "Rejeitadas Uf",
            "type": "integer"
          },
          "rejeitadas_valor": {
            "default": 0,
            "description": "Rejected by value range",
            "title": "Rejeitadas Valor",
            "type": "integer"
          }
        },
        "title": "FilterStats",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StatusLicitacao": {
        "description": "Status da licita\u00e7\u00e3o para filtro de busca.\n\nDefine o est\u00e1gio atual do processo licitat\u00f3rio:\n- RECEBENDO_PROPOSTA: Licita\u00e7\u00f5es abertas para envio de propostas (padr\u00e3o)\n- EM_JULGAMENTO: Propostas encerradas, em an\u00e1lise pelo \u00f3rg\u00e3o\n- ENCERRADA: Processo finalizado (com ou sem vencedor)\n- TODOS: Sem filtro de status (retorna todas)",
        "enum": [
          "recebendo_proposta",
          "em_julgamento",
          "encerrada",
          "todos"
        ],
        "title": "StatusLicitacao",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API para busca e an\u00e1lise de licita\u00e7\u00f5es de uniformes no PNCP. Permite filtrar oportunidades por estado, valor e keywords, gerando relat\u00f3rios Excel e resumos executivos via IA.",
    "title": "BidIQ Uniformes API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/buscar": {
      "post": {
        "description": "Main search endpoint \u2014 thin wrapper that delegates to SearchPipeline.\n\nSTORY-216 AC5: buscar_licitacoes() is a thin wrapper that creates\nSearchPipeline and runs stages. All business logic lives in search_pipeline.py.\n\nThe wrapper handles:\n- SSE tracker lifecycle (create, cleanup on error)\n- Exception mapping (PNCPRateLimitError \u2192 503, PNCPAPIError \u2192 502, etc.)\n- Dependency injection (passing module-level names for test mock compatibility)",
        "operationId": "buscar_licitacoes_buscar_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuscaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuscaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Buscar Licitacoes",
        "tags": [
          "search"
        ]
      }
    },
    "/buscar-progress/{search_id}": {
      "get": {
        "description": "SSE endpoint for real-time search progress updates.\n\nThe client opens this connection simultaneously with POST /buscar,\nusing the same search_id to correlate progress events.\n\nEvents:\n    - connecting (5%): Initial setup\n    - fetching (10-55%): Per-UF progress with uf_index/uf_total\n    - filtering (60-70%): Filter application\n    - llm (75-90%): LLM summary generation\n    - excel (92-98%): Excel report generation\n    - complete (100%): Search finished\n    - error: Search failed",
        "operationId": "buscar_progress_stream_buscar_progress__search_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "search_id",
            "required": true,
            "schema": {
              "title": "Search Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Buscar Progress Stream",
        "tags": [
          "search"
        ]
      }
    },
...
    "/v1/buscar": {
      "post": {
        "description": "Main search endpoint \u2014 thin wrapper that delegates to SearchPipeline.\n\nSTORY-216 AC5: buscar_licitacoes() is a thin wrapper that creates\nSearchPipeline and runs stages. All business logic lives in search_pipeline.py.\n\nThe wrapper handles:\n- SSE tracker lifecycle (create, cleanup on error)\n- Exception mapping (PNCPRateLimitError \u2192 503, PNCPAPIError \u2192 502, etc.)\n- Dependency injection (passing module-level names for test mock compatibility)",
        "operationId": "buscar_licitacoes_v1_buscar_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuscaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuscaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Buscar Licitacoes",
        "tags": [
          "search"
        ]
      }
    },
    "/v1/buscar-progress/{search_id}": {
      "get": {
        "description": "SSE endpoint for real-time search progress updates.\n\nThe client opens this connection simultaneously with POST /buscar,\nusing the same search_id to correlate progress events.\n\nEvents:\n    - connecting (5%): Initial setup\n    - fetching (10-55%): Per-UF progress with uf_index/uf_total\n    - filtering (60-70%): Filter application\n    - llm (75-90%): LLM summary generation\n    - excel (92-98%): Excel report generation\n    - complete (100%): Search finished\n    - error: Search failed",
        "operationId": "buscar_progress_stream_v1_buscar_progress__search_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "search_id",
            "required": true,
            "schema": {
              "title": "Search Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Buscar Progress Stream",
        "tags": [
          "search"
        ]
      }
    },
...
  }
}